+++ Changes.md
@@ -6,6 +6,7 @@ Fixes
 
 - UIEditor : Fixed Python 3 compatibility in plug presets editor.
 - NodeGadget : Fixed intermittent shutdown crash.
+- Expression : Fixed handling of expressions which assign CompoundData to `Attributes.extraAttributes`.
 
 API
 ---